A huge number of changes
- Engine rebooting logic has been overhauled, I would honestly consider
it to be really robust at this point
- Added a new haptic pattern called "Crash" which plays a strong
transient, and then a low rumble that fades out. I intend to use it when
entering a map, and when failing a map. I might also use it when the
game starts
- Fixed bug where slider haptics were locked to 0.3f, and put a (very)
visible warning to try and prevent others from wasting as much time as I
did to stupid apple api spec